Map a Java long to an Objective-C long long.
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1245 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/xmlvm2objc/compat-lib/objc/xmlvm.h
@@ -44,11 +44,11 @@
 void xmlvm_init();
 
 typedef union {
-    id      o;
-    int     i;
-    float   f;
-    double  d;
-    long    l;
+    id        o;
+    int       i;
+    float     f;
+    double    d;
+    long long l;
 } XMLVMElem;
 
 typedef union {
@@ -59,8 +59,8 @@ typedef union {
     int*            i;
     float*          f;
     double*         d;
-    long*           l;
-      void*           data;
+    long long*      l;
+    void*           data;
 } XMLVMElemPtr;
